fix: disable esmoduleInterop setting
This patch disables the `esmoduleInterop` setting that causes type
issues when the library is used by applications that do not transform
their code in this way.

Note the workaround for the `strategy.ts` file (import = require()) is
not ideal, but the export in that file is not a valid ES export, so
using this TS workaround to get past it.

fixes: #482